+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Bhaptics.Tac 1.1.1 (Nov 9th, 2017.)
+* Add new position type (HandL, HandR, FootL, FootR)
+
 # Bhaptics.Tac 1.1.0 (Nov 3th, 2017.)
 * Update websocket communcation between bHaptics Player
 * The version of bHaptics Player should be higher or eqaul to 1.2.0