+++ samples/tac-sharp-unity/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.1
+* Add new position type (HandL, HandR, FootL, FootR)
+
 ## 1.1.0
 * Update websocket communcation between bHaptics Player
 * The version of bHaptics Player should be higher or eqaul to 1.2.0